Perf:Finished sprite_test with succesful use of psram to load sprites
