+++ sdkconfig
@@ -2200,7 +2200,7 @@ CONFIG_LV_USE_BUILTIN_SPRINTF=y
 #
 # HAL Settings
 #
-CONFIG_LV_DEF_REFR_PERIOD=50
+CONFIG_LV_DEF_REFR_PERIOD=33
 CONFIG_LV_DPI_DEF=156
 # end of HAL Settings
